+++ PlayniteVndbExtension/VndbMetadataSettingsView.xaml.cs
@@ -7,11 +7,4 @@
             InitializeComponent();
         }
     }
-
-    public enum SpoilerLevel
-    {
-        None,
-        Minor,
-        Major
-    }
 }